Bug 15395: Fix output of msgcat
By default, msgcat tries to merge translations when two or more files have the
same msgid, even for the header (msgid ""). This resulted in an invalid PO file
("#-#-#-#-#" markers in header).
This patch tells msgcat to use the first available translation instead.

Also, this patch set correctly the Project-Id-Version field

Signed-off-by: Marc Véron <veron@veron.ch>