Bug 15395: Make Koha::I18N work with Plack + add tests
Language detection was done in a BEGIN block so it was not working correctly
with Plack. This is now done when the first translation is requested.

This patch also fixes a bug when set_locale(LC_ALL, ...) fails for some reasons
(locale is not installed system-wide for example).

And finally, add some unit tests

Test plan:
1. prove t/Koha/I18N.t
2. Replay the previous test plans with plack enabled

Signed-off-by: Marc Véron <veron@veron.ch>